Add short-circuiting support to AND and OR
AND and OR expressions now short-circuit. That is, they'll still sometimes return values even if one
side of the expression is erroring (because of a missing simvar or L: variable, for instance).

Specifically:
`X AND Y`  is false if either X or Y are false, irrespective of whether the other value is an error
`X OR Y` is true if either X or Y are true, irrespective of whether the other value is an error

So, the expression `L:VAR1 == 42 OR L:VAR2 == 42` is `true` if `L:VAR1` is `42`, whether or not `L:VAR2` exists.