Fix simvar unsubscribe code
So, despite lots of web pages stating that the way to unsubscribe from simvars is to call
`ClearClientDataDefinition`, this doesn't actually work. This results in us receiving
unexpected values despite having 'unsubscribed'. The correct way is to resubscribe
with a period of `NEVER`.

So, we now simply keep definition ids around so that they can be reused when changing
from one aircraft to another, unsubscribing from a variable and resubscribing to it again.