Allow installation as administrator
When running install.bat/uninstall.bat with Administrator
privileges (which isn't necessary!) the CWD is not
the location of the .bat file, so we tweak the .bat file
to run BBL from the correct relative path.

We also warn the user that running with elevated
privileges isn't necessary. They shouldn't be doing
that as a matter of course; it invites security issues.